refactor: 52개의 카드 생성을 위한 CardFactory 작성
