+++ src/main/java/blackjack/domain/card/Card.java
@@ -1,50 +1,23 @@
 package blackjack.domain.card;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 import java.util.Objects;
 
 public class Card {
-    public static final String NULL_ERR_MSG = "생성자에 Null 이 들어올 수 없습니다.";
-    private static final String INVALID_CARD_TYPE_OR_FIGURE_ERR_MSG = "존재하지 않는 카드 패턴입니다.";
-    private static final List<Card> cards;
+    public static final String NULL_ERR_MSG = "인자에 Null 이 들어올 수 없습니다.";
 
     private Type type;
     private Figure figure;
 
-    static {
-        cards = new ArrayList<>();
-        for (Type type : Type.values()) {
-            for (Figure figure : Figure.values()) {
-                cards.add(new Card(type, figure));
-            }
-        }
-    }
-
     private Card(Type type, Figure figure) {
-        Objects.requireNonNull(type, NULL_ERR_MSG);
-        Objects.requireNonNull(figure, NULL_ERR_MSG);
-
         this.type = type;
         this.figure = figure;
-    }
-
-    public static List<Card> create() {
-        Collections.shuffle(cards);
-        return Collections.unmodifiableList(cards);
     }
 
     public static Card of(Type type, Figure figure) {
         Objects.requireNonNull(type, NULL_ERR_MSG);
         Objects.requireNonNull(figure, NULL_ERR_MSG);
 
-        Card expectedCard = new Card(type, figure);
-
-        return cards.stream()
-                .filter(expectedCard::equals)
-                .findFirst()
-                .orElseThrow(() -> new IllegalArgumentException(INVALID_CARD_TYPE_OR_FIGURE_ERR_MSG));
+        return new Card(type, figure);
     }
 
     public boolean isAce() {
