+++ src/main/java/blackjack/domain/card/CardDeck.java
@@ -10,7 +10,7 @@ public class CardDeck {
     private final Queue<Card> cards;
 
     public CardDeck() {
-        cards = new LinkedList<>(Card.create());
+        cards = new LinkedList<>(CardFactory.create());
     }
 
     public void dealFirstCards(Participant participant) {
