+++ src/test/java/blackjack/domain/card/CardTest.java
@@ -1,37 +1,13 @@
 package blackjack.domain.card;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-
-import java.util.List;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 
 class CardTest {
-
-    @DisplayName("카드 52개 생성 확인")
-    @Test
-    void test1() {
-        int actualCardDeckSize = Card.create().size();
-
-        assertThat(actualCardDeckSize).isEqualTo(52);
-    }
-
-    @DisplayName("52종류 카드 객체 1회 생성 확인")
-    @Test
-    void test2() {
-        List<Card> cards = Card.create();
-        List<Card> expectedCards = Card.create();
-
-        Card card = Card.of(Type.TWO, Figure.HEART);
-        Card expectedCard = Card.of(Type.TWO, Figure.HEART);
-
-        assertThat(card == expectedCard).isTrue();
-        assertThat(cards).isEqualTo(expectedCards);
-    }
 
     @DisplayName("에이스 카드 확인")
     @ParameterizedTest
